requirements.txt
celery==5.3.4
Django==5.0.2
django-celery-results==2.5.1
django-ckeditor-5==0.2.11
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-filter==23.5
django-jazzmin==2.6.0
django-parler==2.3
django-parler-rest==2.2
django-redis==5.4.0
djangorestframework==3.14.0
drf-yasg==1.21.7
ffmpeg-python==0.2.0
pillow==10.2.0
psycopg2-binary==2.9.9
redis==4.6.0
